Update protos for feegrant and authz
