+++ cosmos-sdk-proto-althea/src/prost/TENDERMINT_COMMIT
@@ -1,1 +1,1 @@
-v0.34.13
+v0.34.16
