+++ cosmos-sdk-proto-althea/src/prost/cosmos.base.abci.v1beta1.rs
@@ -41,6 +41,14 @@ pub struct TxResponse {
     /// it's genesis time.
     #[prost(string, tag = "12")]
     pub timestamp: ::prost::alloc::string::String,
+    /// Events defines all the events emitted by processing a transaction. Note,
+    /// these events include those emitted by processing all the messages and those
+    /// emitted from the ante handler. Whereas Logs contains the events, with
+    /// additional metadata, emitted only by processing the messages.
+    ///
+    /// Since: cosmos-sdk 0.42.11, 0.44.5, 0.45
+    #[prost(message, repeated, tag = "13")]
+    pub events: ::prost::alloc::vec::Vec<crate::tendermint::abci::Event>,
 }
 /// ABCIMessageLog defines a structure containing an indexed tx ABCI message log.
 #[derive(Clone, PartialEq, ::prost::Message)]
@@ -143,87 +151,4 @@ pub struct SearchTxsResult {
     /// List of txs in current page
     #[prost(message, repeated, tag = "6")]
     pub txs: ::prost::alloc::vec::Vec<TxResponse>,
-}
-#[cfg(feature = "grpc")]
-#[cfg_attr(docsrs, doc(cfg(feature = "grpc")))]
-#[doc = r" Generated client implementations."]
-pub mod application_query_service_client {
-    #![allow(unused_variables, dead_code, missing_docs, clippy::let_unit_value)]
-    use tonic::codegen::*;
-    #[doc = " ApplicationQueryService exposes the ABCIApplication Query method on BaseApp"]
-    #[derive(Debug, Clone)]
-    pub struct ApplicationQueryServiceClient<T> {
-        inner: tonic::client::Grpc<T>,
-    }
-    impl ApplicationQueryServiceClient<tonic::transport::Channel> {
-        #[doc = r" Attempt to create a new client by connecting to a given endpoint."]
-        pub async fn connect<D>(dst: D) -> Result<Self, tonic::transport::Error>
-        where
-            D: std::convert::TryInto<tonic::transport::Endpoint>,
-            D::Error: Into<StdError>,
-        {
-            let conn = tonic::transport::Endpoint::new(dst)?.connect().await?;
-            Ok(Self::new(conn))
-        }
-    }
-    impl<T> ApplicationQueryServiceClient<T>
-    where
-        T: tonic::client::GrpcService<tonic::body::BoxBody>,
-        T::ResponseBody: Body + Send + 'static,
-        T::Error: Into<StdError>,
-        <T::ResponseBody as Body>::Error: Into<StdError> + Send,
-    {
-        pub fn new(inner: T) -> Self {
-            let inner = tonic::client::Grpc::new(inner);
-            Self { inner }
-        }
-        pub fn with_interceptor<F>(
-            inner: T,
-            interceptor: F,
-        ) -> ApplicationQueryServiceClient<InterceptedService<T, F>>
-        where
-            F: tonic::service::Interceptor,
-            T: tonic::codegen::Service<
-                http::Request<tonic::body::BoxBody>,
-                Response = http::Response<
-                    <T as tonic::client::GrpcService<tonic::body::BoxBody>>::ResponseBody,
-                >,
-            >,
-            <T as tonic::codegen::Service<http::Request<tonic::body::BoxBody>>>::Error:
-                Into<StdError> + Send + Sync,
-        {
-            ApplicationQueryServiceClient::new(InterceptedService::new(inner, interceptor))
-        }
-        #[doc = r" Compress requests with `gzip`."]
-        #[doc = r""]
-        #[doc = r" This requires the server to support it otherwise it might respond with an"]
-        #[doc = r" error."]
-        pub fn send_gzip(mut self) -> Self {
-            self.inner = self.inner.send_gzip();
-            self
-        }
-        #[doc = r" Enable decompressing responses with `gzip`."]
-        pub fn accept_gzip(mut self) -> Self {
-            self.inner = self.inner.accept_gzip();
-            self
-        }
-        #[doc = " Query exposes the ABCIApplication Query method on BaseApp"]
-        pub async fn query(
-            &mut self,
-            request: impl tonic::IntoRequest<crate::tendermint::abci::RequestQuery>,
-        ) -> Result<tonic::Response<crate::tendermint::abci::ResponseQuery>, tonic::Status>
-        {
-            self.inner.ready().await.map_err(|e| {
-                tonic::Status::new(
-                    tonic::Code::Unknown,
-                    format!("Service was not ready: {}", e.into()),
-                )
-            })?;
-            let codec = tonic::codec::ProstCodec::default();
-            let path = http::uri::PathAndQuery::from_static(
-                "/cosmos.base.abci.v1beta1.ApplicationQueryService/Query",
-            );
-            self.inner.unary(request.into_request(), path, codec).await
-        }
-    }
 }
