+++ cosmos-sdk-proto-althea/src/prost/cosmos.base.tendermint.v1beta1.rs
@@ -110,6 +110,7 @@ pub struct VersionInfo {
     pub go_version: ::prost::alloc::string::String,
     #[prost(message, repeated, tag = "7")]
     pub build_deps: ::prost::alloc::vec::Vec<Module>,
+    /// Since: cosmos-sdk 0.43
     #[prost(string, tag = "8")]
     pub cosmos_sdk_version: ::prost::alloc::string::String,
 }
