+++ cosmos-sdk-proto-althea/src/prost/cosmos.vesting.v1beta1.rs
@@ -151,6 +151,8 @@ pub struct PeriodicVestingAccount {
 /// PermanentLockedAccount implements the VestingAccount interface. It does
 /// not ever release coins, locking them indefinitely. Coins in this account can
 /// still be used for delegating and for governance votes even while locked.
+///
+/// Since: cosmos-sdk 0.43
 #[derive(Clone, PartialEq, ::prost::Message)]
 pub struct PermanentLockedAccount {
     #[prost(message, optional, tag = "1")]
