+++ cosmos-sdk-proto-althea/src/prost/ibc.applications.transfer.v1.rs
@@ -1,21 +1,3 @@
-/// FungibleTokenPacketData defines a struct for the packet payload
-/// See FungibleTokenPacketData spec:
-/// <https://github.com/cosmos/ics/tree/master/spec/ics-020-fungible-token-transfer#data-structures>
-#[derive(Clone, PartialEq, ::prost::Message)]
-pub struct FungibleTokenPacketData {
-    /// the token denomination to be transferred
-    #[prost(string, tag = "1")]
-    pub denom: ::prost::alloc::string::String,
-    /// the token amount to be transferred
-    #[prost(uint64, tag = "2")]
-    pub amount: u64,
-    /// the sender address
-    #[prost(string, tag = "3")]
-    pub sender: ::prost::alloc::string::String,
-    /// the recipient address on the destination chain
-    #[prost(string, tag = "4")]
-    pub receiver: ::prost::alloc::string::String,
-}
 /// DenomTrace contains the base denomination for ICS20 fungible tokens and the
 /// source tracing information path.
 #[derive(Clone, PartialEq, ::prost::Message)]
@@ -219,7 +201,7 @@ pub mod query_client {
 }
 /// MsgTransfer defines a msg to transfer fungible tokens (i.e Coins) between
 /// ICS20 enabled chains. See ICS Spec here:
-/// <https://github.com/cosmos/ics/tree/master/spec/ics-020-fungible-token-transfer#data-structures>
+/// <https://github.com/cosmos/ibc/tree/master/spec/app/ics-020-fungible-token-transfer#data-structures>
 #[derive(Clone, PartialEq, ::prost::Message)]
 pub struct MsgTransfer {
     /// the port on which the packet will be sent
