+++ cosmos-sdk-proto-althea/src/prost/ibc.applications.transfer.v2.rs
@@ -1,0 +1,18 @@
+/// FungibleTokenPacketData defines a struct for the packet payload
+/// See FungibleTokenPacketData spec:
+/// <https://github.com/cosmos/ibc/tree/master/spec/app/ics-020-fungible-token-transfer#data-structures>
+#[derive(Clone, PartialEq, ::prost::Message)]
+pub struct FungibleTokenPacketData {
+    /// the token denomination to be transferred
+    #[prost(string, tag = "1")]
+    pub denom: ::prost::alloc::string::String,
+    /// the token amount to be transferred
+    #[prost(string, tag = "2")]
+    pub amount: ::prost::alloc::string::String,
+    /// the sender address
+    #[prost(string, tag = "3")]
+    pub sender: ::prost::alloc::string::String,
+    /// the recipient address on the destination chain
+    #[prost(string, tag = "4")]
+    pub receiver: ::prost::alloc::string::String,
+}
